pyproject.toml
[project]
name = "django-gpg-keys-manager"
version = "0.1.0"
description = "Django reusable app to manage GPG public keys."
readme = "README.md"
authors = [{ name = "Laurent FAVOLE", email = "laurentfavole03@gmail.com" }]
license = "Unlicense"
license-files = ["LICENSE"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Sound/Audio",
]
requires-python = ">=3.13"
keywords = ["django", "gpg"]
dependencies = [
    "Django>=5.0,<7.0",
    "pgpy13~=0.6",
]

[tool.vercel.scripts]
build = "bash build.sh"

[project.optional-dependencies]
dev = ["bump-my-version", "ruff"]

[project.urls]
Homepage = "https://github.com/lfavole/django-gpg-keys-manager"

[tool.bumpversion]
current_version = "0.1.0"
message = "chore: Release {new_version}"
commit = true
search = "{current_version}"
replace = "{new_version}"
regex = false
sign_tags = true
tag = true
tag_name = "{new_version}"
tag_message = "{new_version}"

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = "version = \"{current_version}\""
replace = "version = \"{new_version}\""

[[tool.bumpversion.files]]
filename = "gpg_keys/__init__.py"
search = "__version__ = \"{current_version}\""
replace = "__version__ = \"{new_version}\""

[tool.ruff]
line-length = 120

[tool.ruff.lint]
preview = true
select = ["ALL"]
ignore = ["ANN002", "ANN003", "COM812", "CPY", "D203", "D212", "EXE"]
exclude = ["gpg_keys/migrations/*.py"]

[tool.ruff.format]
preview = true

[tool.setuptools]
packages = ["gpg_keys"]
